feat: add voice note

+++ app/src/main/java/dev/dizyaa/dizgram/feature/chat/domain/VoiceNote.kt
@@ -3,5 +3,5 @@ package dev.dizyaa.dizgram.feature.chat.domain
 data class VoiceNote(
     val file: File,
     val duration: Int,
-    val waveform: List<Byte>,
+    val waveform: List<Int>,
 )
